requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dataclass-wizard==0.22.2
decorator==4.4.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.2
opencv-python==4.9.0.80
phonenumbers==8.13.27
Pillow==10.1.0
proglog==0.1.10
requests==2.31.0
tqdm==4.66.1
urllib3==2.1.0
